requirements.txt
asgiref==3.5.2
Django==4.1.2
django-cleanup==6.0.0
django-phonenumber-field==7.0.0
django-phonenumbers==1.0.1
django-widget-tweaks==1.4.12
phonenumbers==8.12.57
Pillow==9.2.0
sqlparse==0.4.3
tzdata==2022.5
